feat: move all query handler into base datasource
