+++ packages/BaseDatasource/src/BaseCollection.php
@@ -5,11 +5,11 @@ namespace ForestAdmin\AgentPHP\BaseDatasource;
 use Doctrine\DBAL\Schema\Column;
 use Doctrine\DBAL\Schema\Table;
 use ForestAdmin\AgentPHP\Agent\Utils\ForestSchema\FrontendFilterable;
-use ForestAdmin\AgentPHP\Agent\Utils\QueryAggregate;
-use ForestAdmin\AgentPHP\Agent\Utils\QueryConverter;
 use ForestAdmin\AgentPHP\BaseDatasource\Contracts\BaseCollectionContract;
 use ForestAdmin\AgentPHP\BaseDatasource\Contracts\BaseDatasourceContract;
 use ForestAdmin\AgentPHP\BaseDatasource\Utils\DataTypes;
+use ForestAdmin\AgentPHP\BaseDatasource\Utils\QueryAggregate;
+use ForestAdmin\AgentPHP\BaseDatasource\Utils\QueryConverter;
 use ForestAdmin\AgentPHP\DatasourceToolkit\Collection as ForestCollection;
 use ForestAdmin\AgentPHP\DatasourceToolkit\Components\Caller;
 use ForestAdmin\AgentPHP\DatasourceToolkit\Components\Query\Aggregation;
