+++ packages/BaseDatasource/src/Utils/QueryBuilder.php
@@ -43,23 +43,25 @@ class QueryBuilder
     public function formatField(string $field): string
     {
         if (Str::contains($field, ':')) {
-            $relation = $this->collection->getFields()[Str::before($field, ':')];
+            $relationName = Str::before($field, ':');
+            $relation = $this->collection->getFields()[$relationName];
             $tableName = $this->collection
                 ->getDataSource()
                 ->getCollection($relation->getForeignCollection())->getTableName();
-            $this->addJoinRelation($relation, $tableName);
+            $this->addJoinRelation($relation, $tableName, $relationName);
 
-            return $tableName . '.' . Str::after($field, ':');
+            return Str::replace(':', '.', $field);
         }
 
         return $this->tableName . '.' . $field;
     }
 
-    protected function addJoinRelation(RelationSchema $relation, string $relationTableName): void
+    protected function addJoinRelation(RelationSchema $relation, string $relationTableName, ?string $relationTableAlias = null): void
     {
+        $relationTableAlias = $relationTableAlias ?? $relationTableName;
         if ($relation instanceof ManyToManySchema) {
             $throughTable = $this->collection->getDataSource()->getCollection($relation->getThroughCollection())->getTableName();
-            $joinTable = "$relationTableName as $relationTableName";
+            $joinTable = "$relationTableName as $relationTableAlias";
             $joinThroughTable = "$throughTable as $throughTable";
             if (! $this->isJoin($joinTable) && ! $this->isJoin($joinThroughTable)) {
                 $this->query
@@ -70,14 +72,14 @@ class QueryBuilder
                         $throughTable . '.' . $relation->getOriginKey()
                     )
                     ->leftJoin(
-                        "$relationTableName as $relationTableName",
+                        "$relationTableName as $relationTableAlias",
                         $throughTable . '.' . $relation->getForeignKey(),
                         '=',
-                        $relationTableName . '.' . $relation->getForeignKeyTarget()
+                        $relationTableAlias . '.' . $relation->getForeignKeyTarget()
                     );
             }
         } else {
-            $joinTable = "$relationTableName as $relationTableName";
+            $joinTable = "$relationTableName as $relationTableAlias";
             if (
                 ($relation instanceof OneToOneSchema || $relation instanceof OneToManySchema)
                 && ! $this->isJoin($joinTable)
@@ -86,14 +88,14 @@ class QueryBuilder
                     $joinTable,
                     $this->tableName . '.' . $relation->getOriginKey(),
                     '=',
-                    $relationTableName . '.' . $relation->getOriginKeyTarget()
+                    $relationTableAlias . '.' . $relation->getOriginKeyTarget()
                 );
             } elseif ($relation instanceof ManyToOneSchema && ! $this->isJoin($joinTable)) {
                 $this->query->leftJoin(
                     $joinTable,
                     $this->tableName . '.' . $relation->getForeignKey(),
                     '=',
-                    $relationTableName . '.' . $relation->getForeignKeyTarget()
+                    $relationTableAlias . '.' . $relation->getForeignKeyTarget()
                 );
             }
         }
